incident
Added:
 - repo for the incident class
 - injected it into the module

Update:
 - using the my Result class

+++ app/src/main/java/com/example/myapplication/features/incident/data/network/model/ApiIncidentItem.kt
@@ -15,9 +15,9 @@ data class ApiIncidentItem(
     @SerializedName("issuerId")
     val issuerId: String?,
     @SerializedName("latitude")
-    val latitude: Int?,
+    val latitude: Double?,
     @SerializedName("longitude")
-    val longitude: Int?,
+    val longitude: Double?,
     @SerializedName("medias")
     val medias: List<ApiMediaItem>?,
     @SerializedName("priority")
